# nbi/swagger_server/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "REST-API server for the North Bound Interface (NBI) of the 5GT-SO.\
    \ Find more at http://5g-transformer.eu"
  version: "1.5"
  title: "5GT-MTP NBI"
  termsOfService: "/tos"
  contact:
    email: "cnd@lists.cttc.es"
host: "localhost:8090"
basePath: "/5gt/mtp/v1"
paths:
  /abstract-resources:
    get:
      tags:
      - "abstractResources"
      summary: "Retrieve aggregated Cloud NFVI-PoP and Inter-NFVI-PoP Connectivity"
      description: "Retrieve aggregated Cloud NFVI-PoP and Inter-NFVI-PoP Connectivity"
      operationId: "collect_mtp_cloud_network_resource_information"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "Successful operation"
          schema:
            type: "object"
            properties:
              NfviPops:
                $ref: "#/definitions/NfviPops"
              logicalLinkInterNfviPops:
                $ref: "#/definitions/logicalLinkInterNfviPops"
        400:
          description: "Bad request"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
      x-swagger-router-controller: "swagger_server.controllers.abstract_resources_controller"
  /abstract-federated-resources:
    get:
      tags:
      - "abstractResources"
      summary: "Retrieve aggregated Cloud NFVI-PoP and Inter-NFVI-PoP Connectivity\
        \ for Federation"
      description: "Retrieve aggregated Cloud NFVI-PoP and Inter-NFVI-PoP Connectivity\
        \ for Federation"
      operationId: "collect_mtp_federated_cloud_network_resource_information"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "Successful operation"
          schema:
            type: "object"
            properties:
              NfviPops:
                $ref: "#/definitions/NfviPops"
              logicalLinkInterNfviPops:
                $ref: "#/definitions/logicalLinkInterNfviPops"
        400:
          description: "Bad request"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
      x-swagger-router-controller: "swagger_server.controllers.abstract_resources_controller"
  /abstract-network-resources:
    post:
      tags:
      - "abstractResources"
      summary: "Create inter-NFVI-PoP connectivity"
      description: ""
      operationId: "create_inter_nfvi_po_p_connectivity"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Create inter-NfviPop Connectivity"
        required: true
        schema:
          $ref: "#/definitions/interNfviPopConnectivityRequest"
      responses:
        201:
          description: "Successful operation"
          schema:
            type: "array"
            items:
              type: "object"
              properties:
                interNfviPopConnnectivityId:
                  $ref: "#/definitions/interNfviPopConnnectivityId"
                interNfviPopNetworkSegmentType:
                  $ref: "#/definitions/interNfviPopNetworkSegmentType"
              required:
              - "interNfviPopConnnectivityId"
              - "interNfviPopNetworkSegmentType"
        400:
          description: "Bad request"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
      x-swagger-router-controller: "swagger_server.controllers.abstract_resources_controller"
    delete:
      tags:
      - "abstractResources"
      summary: "Delete inter-NFVI-PoP connectivity"
      description: ""
      operationId: "delete_inter_nfvi_po_p_connectivity"
      consumes:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Delete inter-NfviPop Connectivity"
        required: true
        schema:
          $ref: "#/definitions/deleteInterNfviPopConnectivityRequest"
      responses:
        204:
          description: "Successful operation"
        400:
          description: "Bad request"
        401:
          description: "Unauthorized"
        403:
          description: "Forbidden"
      x-swagger-router-controller: "swagger_server.controllers.abstract_resources_controller"
...
definitions:
  NfviPops:
    type: "array"
    items:
      $ref: "#/definitions/NfviPops_inner"
  networkLayer:
    type: "string"
    description: "5GT - Describes the network layer capabilities of the Gw Layer 2\
      \ (L2) or Layer 3 (L3), i.e. Ethernet, IP, IP/MPLS -- see ETSI NFV-IFA 022 6.4.5.3."
  interNfviPopNetworkType:
    type: "string"
    description: "5GT - Describes the network connectivity types, e.g., L2-VPN, EoMPLS,\
      \ VXLAN, MPLS over GRE, ... -- see ETSI NFV-IFA 022 6.4.5.3."
  interNfviPopNetworkTopology:
    type: "string"
    description: "5GT - Describes the network connectivity topology, i.e., p2p, mesh,\
      \ tree -- see ETSI NFV-IFA 022 6.4.5.2."
  logicalLinkInterNfviPops:
    type: "array"
    items:
      $ref: "#/definitions/logicalLinkInterNfviPops_inner"
  interNfviPopConnectivityRequest:
    type: "object"
    required:
    - "interNfviPopNetworkType"
    - "logicalLinkPathList"
    - "metaData"
    - "networkLayer"
    properties:
      logicalLinkPathList:
        $ref: "#/definitions/logicalLinkPathList"
      networkLayer:
        $ref: "#/definitions/networkLayer"
      interNfviPopNetworkType:
        $ref: "#/definitions/interNfviPopNetworkType"
      metaData:
        $ref: "#/definitions/metaData"
  interNfviPopConnnectivityId:
    type: "string"
    description: "Identifier provided by the MTP to identify the inter-Nfvi-Pop connectivity."
  metaData:
    type: "array"
    description: "metaData information used to carry ionformat metadata."
    items:
      $ref: "#/definitions/metaData_inner"
  interNfviPopNetworkSegmentType:
    type: "string"
    description: "5GT - provides the selected segment id, e.g., VLAN Id, VXLAN Id,\
      \ ... -- see ETSI NFV-IFA 022 6.4.5.2."
  deleteInterNfviPopConnectivityRequest:
    type: "object"
    required:
    - "interNfviPopConnnectivityIdList"
    - "metaData"
    properties:
      interNfviPopConnnectivityIdList:
        $ref: "#/definitions/interNfviPopConnnectivityIdList"
      metaData:
        $ref: "#/definitions/metaData"
  logicalLinkAttributes:
    type: "object"
    required:
    - "dstGwIpAddress"
    - "localLinkId"
    - "logicalLinkId"
    - "remoteLinkId"
    - "srcGwIpAddress"
    properties:
      logicalLinkId:
        type: "string"
        description: "(numbered) Identifier of the logical link"
      srcGwIpAddress:
        type: "string"
        description: "5GT - Source NFVI-PoP Gw IPv4 Address in terms of A.B.C.D (/32)."
      localLinkId:
        type: "integer"
        format: "uint32"
        description: "Local Logical Link Id."
      dstGwIpAddress:
        type: "string"
        description: "5GT - Destination NFVI-PoP Gw IPv4 Address in terms of A.B.C.D\
          \ (/32)."
      remoteLinkId:
        type: "integer"
        format: "uint32"
        description: "Remote Logical Link Id."
    description: "5GT - inter-Nfvi-Pop connectivity link."
  logicalLinkPathList:
    type: "array"
    items:
      $ref: "#/definitions/logicalLinkPathList_inner"
  interNfviPopConnnectivityIdList:
    type: "array"
    items:
      $ref: "#/definitions/interNfviPopConnnectivityIdList_inner"
...
  NfviPops_inner_nfviPopAttributes_networkConnectivityEndpoint:
    required:
    - "netGwIpAddress"
    properties:
      netGwIpAddress:
        type: "string"
        description: "5GT - Reachable Gw IPv4 Address in terms of A.B.C.D (/32)."
    description: "Information about network connectivity endpoints to the NFVI-PoP\
      \ that the VIM manages which helps build topology information relative to NFVI-PoP\
      \ connectivity to other NFVI-PoP or N-PoP."
  NfviPops_inner_nfviPopAttributes_memoryResourceAttributes:
    required:
    - "allocatedCapacity"
    - "availableCapacity"
    - "reservedCapacity"
    - "totalCapacity"
    properties:
      availableCapacity:
        type: "string"
        description: "The free capacity available for allocation and reservation of\
          \ memory resources."
      reservedCapacity:
        type: "string"
        description: "The reserved capacity of   memory resources."
      totalCapacity:
        type: "string"
        description: "The total capacity of memory resources."
      allocatedCapacity:
        type: "string"
        description: "The allocated capacity of memory resources."
  NfviPops_inner_nfviPopAttributes_cpuResourceAttributes:
    required:
    - "allocatedCapacity"
    - "availableCapacity"
    - "reservedCapacity"
    - "totalCapacity"
    properties:
      availableCapacity:
        type: "string"
        description: "The free capacity available for allocation and reservation of\
          \ CPU resources."
      reservedCapacity:
        type: "string"
        description: "The reserved capacity of   CPU resources."
      totalCapacity:
        type: "string"
        description: "The total capacity of CPU resources."
      allocatedCapacity:
        type: "string"
        description: "The allocated capacity of CPU resources."
  NfviPops_inner_nfviPopAttributes_storageResourceAttributes:
    required:
    - "allocatedCapacity"
    - "availableCapacity"
    - "reservedCapacity"
    - "totalCapacity"
    properties:
      availableCapacity:
        type: "string"
        description: "The free capacity available for allocation and reservation of\
          \ storage resources."
      reservedCapacity:
        type: "string"
        description: "The reserved capacity of storage resources."
      totalCapacity:
        type: "string"
        description: "The total capacity of storage resources."
      allocatedCapacity:
        type: "string"
        description: "The allocated capacity of storage resources."
  NfviPops_inner_nfviPopAttributes_resourceZoneAttributes:
    required:
    - "cpuResourceAttributes"
    - "memoryResourceAttributes"
    - "metadata"
    - "storageResourceAttributes"
    - "zoneId"
    - "zoneName"
    - "zoneProperty"
    - "zoneState"
    properties:
      zoneId:
        type: "string"
        description: "The identifier of the Resource Zone."
      zoneName:
        type: "string"
        description: "The name of the Resource Zone."
      zoneState:
        type: "string"
        description: "Information about the current state of the Resource Zone, e.g.\
          \ if the Resource Zone is available."
      zoneProperty:
        type: "string"
        description: "Set of properties that define the capabilities associated to\
          \ the Resource Zone. Examples of capabilities may include: support of certain\
          \ compute resource types (e.g. low performance, acceleration capabilities,\
          \ etc. association to certain NFVI-PoP physical segregation (e.g. different\
          \ power or network sub-systems, availability of redundancy power sub-systems),\
          \ etc."
      metadata:
        type: "string"
        description: "List of metadata key-value pairs used to associate meaningful\
          \ metadata to the NFVI-PoP Zone."
      memoryResourceAttributes:
        $ref: "#/definitions/NfviPops_inner_nfviPopAttributes_memoryResourceAttributes"
      cpuResourceAttributes:
        $ref: "#/definitions/NfviPops_inner_nfviPopAttributes_cpuResourceAttributes"
      storageResourceAttributes:
        $ref: "#/definitions/NfviPops_inner_nfviPopAttributes_storageResourceAttributes"
  NfviPops_inner_nfviPopAttributes:
    required:
    - "federatedVimId"
    - "geographicalLocationInfo"
    - "networkConnectivityEndpoint"
    - "nfviPopId"
    - "resourceZoneAttributes"
    - "vimId"
    properties:
      geographicalLocationInfo:
        type: "string"
        description: "It provides information about the geographic location (e.g.\
          \ geographic coordinates or address of the building, etc.) of the NFVI resources\
          \ that the VIM manages."
      vimId:
        type: "string"
        description: "Identification of the VIM."
      federatedVimId:
        type: "string"
        description: "Identification of the Federated VIM."
      networkConnectivityEndpoint:
        type: "array"
        items:
          $ref: "#/definitions/NfviPops_inner_nfviPopAttributes_networkConnectivityEndpoint"
      nfviPopId:
        type: "string"
        description: "Identification of the NFVI-PoP."
      resourceZoneAttributes:
        type: "array"
        items:
          $ref: "#/definitions/NfviPops_inner_nfviPopAttributes_resourceZoneAttributes"
  NfviPops_inner:
    required:
    - "nfviPopAttributes"
    properties:
      nfviPopAttributes:
        $ref: "#/definitions/NfviPops_inner_nfviPopAttributes"
  logicalLinkInterNfviPops_inner_logicalLinks_networkQoS:
    required:
    - "linkCost"
    - "linkCostValue"
    - "linkDelay"
    - "linkDelayValue"
    properties:
      linkCostValue:
        type: "number"
        description: "5GT - Value of Link Cost"
      linkCost:
        type: "string"
        description: "5GT - Link cost name ."
      linkDelayValue:
        type: "number"
        description: "5GT - Value of Link Delay (expressed in ms)"
      linkDelay:
        type: "string"
        description: "5GT - Link delay name."
    description: "5GT - Element providing information about Quality of Service attributes\
      \ (cost and delay) that the network shall support."
  logicalLinkInterNfviPops_inner_logicalLinks:
    required:
    - "availableBandwidth"
    - "dstGwIpAddress"
    - "interNfviPopNetworkTopology"
    - "interNfviPopNetworkType"
    - "localLinkId"
    - "logicalLinkId"
    - "networkLayer"
    - "networkQoS"
    - "remoteLinkId"
    - "srcGwIpAddress"
    - "totalBandwidth"
    properties:
      logicalLinkId:
        type: "string"
        description: "(numbered) Identifier of the logical link"
      totalBandwidth:
        type: "number"
        description: "Total bandwidth capacity supported by the logical link (in Mbps)."
      availableBandwidth:
        type: "number"
        description: "Available bandwidth capacity supported by the logical link (in\
          \ Mbps)."
      networkQoS:
        $ref: "#/definitions/logicalLinkInterNfviPops_inner_logicalLinks_networkQoS"
      srcGwIpAddress:
        type: "string"
        description: "5GT - Source NFVI-PoP Gw IPv4 Address in terms of A.B.C.D (/32)."
      localLinkId:
        type: "integer"
        format: "uint32"
        description: "Local Logical Link Id."
      dstGwIpAddress:
        type: "string"
        description: "5GT - Destination NFVI-PoP Gw IPv4 Address in terms of A.B.C.D\
          \ (/32)."
      remoteLinkId:
        type: "integer"
        format: "uint32"
        description: "Remote Logical Link Id."
      networkLayer:
        $ref: "#/definitions/networkLayer"
      interNfviPopNetworkType:
        $ref: "#/definitions/interNfviPopNetworkType"
      interNfviPopNetworkTopology:
        $ref: "#/definitions/interNfviPopNetworkTopology"
    description: "5GT - Element providing information the attributes related to particular\
      \ logical links between pair of NFVI-PoPs."
  logicalLinkInterNfviPops_inner:
    required:
    - "logicalLinks"
    properties:
      logicalLinks:
        $ref: "#/definitions/logicalLinkInterNfviPops_inner_logicalLinks"
  metaData_inner:
    required:
    - "key"
    - "value"
    properties:
      key:
        type: "string"
      value:
        type: "string"
    example:
      value: "value"
      key: "key"
  logicalLinkPathList_inner:
    required:
    - "logicalLinkAttributes"
    - "reqBandwidth"
    - "reqLatency"
    properties:
      logicalLinkAttributes:
        $ref: "#/definitions/logicalLinkAttributes"
      reqBandwidth:
        type: "number"
        description: "requested bandwidth (in Mbps)."
      reqLatency:
        type: "number"
        description: "5GT - requested maximum end-to-end latency (expressed in ms)"
      metaData:
        $ref: "#/definitions/metaData"
  interNfviPopConnnectivityIdList_inner:
    required:
    - "interNfviPopConnnectivityId"
    properties:
      interNfviPopConnnectivityId:
        $ref: "#/definitions/interNfviPopConnnectivityId"
...